+++ semcore/data-table/src/components/Body/Row.tsx
@@ -337,6 +337,7 @@ export class RowRoot<Data extends DataTableData, UniqKeyType> extends Component<
       shadowVertical,
       variant,
       flatRows,
+      sideIndents,
     } = this.asProps;
 
     const expanded = expandedRows?.has(row[UNIQ_ROW_KEY]) && !this.state.expandedForAnimation;
@@ -510,6 +511,7 @@ export class RowRoot<Data extends DataTableData, UniqKeyType> extends Component<
                   accordionDuration={accordionDuration}
                   variant={variant}
                   flatRows={flatRows}
+                  sideIndents={sideIndents}
                 />
               );
             })}