+++ stories/components/input-tags/tests/examples/tags-with-addons.tsx
@@ -18,6 +18,7 @@ const Demo = (props: ExampleInputTagsProps) => {
             disabled={props.disabled}
             interactive={props.interactive}
             color={props.color}
+            addonRight={Edit}
           >
             <InputTags.Tag.Text>
               <InputTags.Tag.Addon>
@@ -25,6 +26,7 @@ const Demo = (props: ExampleInputTagsProps) => {
               </InputTags.Tag.Addon>
               <InputTags.Tag.Text.Content>Text and addon</InputTags.Tag.Text.Content>
             </InputTags.Tag.Text>
+
           </InputTags.Tag>
 
           <InputTags.Tag
@@ -35,12 +37,11 @@ const Demo = (props: ExampleInputTagsProps) => {
             interactive={props.interactive}
 
             color={props.color}
+            addonLeft={Edit}
+            active
           >
             <InputTags.Tag.Text>
-              <InputTags.Tag.Addon>
-                <Edit />
-              </InputTags.Tag.Addon>
-              <InputTags.Tag.Text.Content>Addon text and close</InputTags.Tag.Text.Content>
+              Addon text and close
             </InputTags.Tag.Text>
             <InputTags.Tag.Close />
           </InputTags.Tag>