+++ semcore/data-table/src/components/Head/Group.tsx
@@ -33,26 +33,8 @@ export class Group extends Component<
   render() {
     const SGroupContainer = Box;
     const SGroup = Root;
-    const { styles, Children, title, fixed, columns, withConfig, getFixedStyle, shadowVertical } = this.asProps;
+    const { styles, Children, title, columns, withConfig } = this.asProps;
     const groupColumns = columns ?? [];
-
-    const firstColumn = groupColumns[0];
-    const lastColumn = groupColumns[groupColumns.length - 1];
-
-    const style: any = {};
-
-    if (fixed === 'left' && firstColumn) {
-      const [name, value] = getFixedStyle({ name: firstColumn.name, fixed: 'left' });
-      if (name !== undefined && value !== undefined) {
-        style[name] = value;
-      }
-    }
-    if (fixed === 'right' && lastColumn) {
-      const [name, value] = getFixedStyle({ name: lastColumn.name, fixed: 'right' });
-      if (name !== undefined && value !== undefined) {
-        style[name] = value;
-      }
-    }
 
     return sstyled(styles)(
       <SGroupContainer data-group-container>
@@ -61,7 +43,6 @@ export class Group extends Component<
           style={style}
           __excludeProps={['title']}
           id={this.groupId}
-          use:shadowVertical={(firstColumn.showShadowVertical || lastColumn.showShadowVertical) ? shadowVertical : undefined}
         >
           {withConfig ? <Children /> : title}
         </SGroup>