+++ semcore/data-table/src/components/DataTable/DataTable.types.ts
@@ -35,7 +35,7 @@ export type DataRowItem = {
 };
 export interface DTValue {
   toString(): string;
-  [ACCORDION]?: React.ReactNode | DataTableData;
+  [ACCORDION]?: React.ReactNode;
 }
 export type DataTableData = DataRowItem[];
 
@@ -91,6 +91,7 @@ export type DataTableProps<
 
     /**
      * Set of expanded rows (uniq id from them)
+     * This is mutable! variable because of table performance. Don't change the link on it.
      */
     expandedRows?: Set<UniqKeyType>;