+++ semcore/data-table/CHANGELOG.md
@@ -8,6 +8,12 @@ CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangel
 
 - Reset expanded accordions if data has changed.
 
+### Fixed
+
+- Render grouped header in some cases.
+- Styles for cells in accordion in different variants.
+- Laggy animation in accordions.
+
 ## [16.3.0] - 2025-09-12
 
 ### Added