+++ semcore/data-table/CHANGELOG.md
@@ -7,6 +7,8 @@ CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangel
 ### Fixed
 
 - Render grouped header in some cases.
+- Styles for cells in accordion in different variants.
+- Laggy animation in accordions.
 
 ## [16.3.0] - 2025-09-12