+++ semcore/data-table/CHANGELOG.md
@@ -9,6 +9,7 @@ CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangel
 - Cell display in the accordion was changed to `block`.
 - Render grouped header in some cases.
 - Styles for cells in accordion in different variants.
+- Laggy animation in accordions.
 
 ## [16.3.0] - 2025-09-12