+++ semcore/data-table/CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangel
 
 ### Fixed
 
+- Cell display in the accordion was changed to `block`.
 - Render grouped header in some cases.
 - Styles for cells in accordion in different variants.
 - Laggy animation in accordions.