+++ semcore/data-table/CHANGELOG.md
@@ -6,10 +6,12 @@ CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangel
 
 ### Fixed
 
+- Cell display in the accordion was changed to `block`.
 - Render table with expanded rows after changing page.
 - Render grouped header in some cases.
 - Styles for cells in accordion in different variants.
 - Laggy animation in accordions.
+- `renderCell` not working for accordion's cells.
 
 ## [16.3.0] - 2025-09-12