+++ semcore/data-table/CHANGELOG.md
@@ -2,16 +2,27 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
-## [16.3.1] - 2025-09-20
+## [16.3.3] - 2025-10-03
+
+### Fixed
+
+- Render table with expanded rows after changing page.
+- Render grouped header in some cases.
+
+## [16.3.2] - 2025-09-20
 
 ### Fixed
 
 - Cell display in the accordion was changed to `block`.
-- Render table with expanded rows after changing page.
-- Render grouped header in some cases.
 - Styles for cells in accordion in different variants.
 - Laggy animation in accordions.
 - `renderCell` not working for accordion's cells.
+
+## [16.3.1] - 2025-09-20
+
+### Fixed
+
+- Render grouped header in some cases.
 
 ## [16.3.0] - 2025-09-12